Modifications so as not to miss days at start of the year
Not doing sorting based on day of week (dow) for the `\crosstabview` psql
meta-command. Sorting on dow will miss some of the days if the dates don't
start with Sunday! Luckily 2023 started on a sunday so this didn't cause
a problem but we can see the issue if we try the year 2021.

Another issue that including the sortcolH (for the crosstabview
meta-command) causes is that the last line of December gets jumbled up!